+++ tests/__init__.py
@@ -25,7 +25,7 @@ from uuid import UUID
 
 from sortedcontainers import SortedSet
 
-from cyclonedx.output import BomRefDiscriminator as _BomRefDiscriminator
+from cyclonedx.contrib.bom.utils import BomRefDiscriminator as _BomRefDiscriminator
 from cyclonedx.schema import OutputFormat, SchemaVersion
 
 if TYPE_CHECKING:
